% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{oh_countries}
\alias{oh_countries}
\title{Get list of countries supported by OpenHolidays}
\usage{
oh_countries(language = "EN")
}
\arguments{
\item{language}{Language code (default: "EN").}
}
\value{
A data frame of supported countries.
}
\description{
Get list of countries supported by OpenHolidays
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oh_countries(language = "DE")
\dontshow{\}) # examplesIf}
}
