% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{oh_languages}
\alias{oh_languages}
\title{Get supported languages}
\usage{
oh_languages()
}
\value{
A data frame of available languages (ISO code + localized name).
}
\description{
This function retrieves the list of supported languages from the
OpenHolidays API.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oh_languages()
\dontshow{\}) # examplesIf}
}
