% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{oh_request}
\alias{oh_request}
\title{Build an OpenHolidays API request}
\usage{
oh_request(endpoint, query = list())
}
\arguments{
\item{endpoint}{API endpoint (e.g. "Countries", "HolidayTypes").}

\item{query}{Named list of query parameters.}
}
\value{
A httr2 request object.
}
\description{
Build an OpenHolidays API request
}
\examples{
req <- oh_request("Countries")
}
