% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional.R
\name{oh_subdivisions}
\alias{oh_subdivisions}
\title{Get subdivisions for a given country}
\usage{
oh_subdivisions(country, language = "EN")
}
\arguments{
\item{country}{Country ISO code (e.g., "DE" for Germany).}

\item{language}{Language code (default: "EN").}
}
\value{
A data frame of subdivisions.
}
\description{
This function retrieves administrative subdivisions (e.g., states, provinces)
for a specified country from the OpenHolidays API.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
oh_subdivisions(country = "DE")
\dontshow{\}) # examplesIf}
}
