% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openholidaysR-package.R
\docType{package}
\name{openholidaysR-package}
\alias{openholidaysR}
\alias{openholidaysR-package}
\title{openholidaysR: Provides access to the OpenHolidays API}
\description{
Provides dates for public and school holidays for a number of countries and their subdivisions through the OpenHolidays API at https://openholidaysapi.org/.
}
\author{
\strong{Maintainer}: Johannes Feldhege \email{mail@johannesfeldhege.de} (\href{https://orcid.org/0000-0002-4665-5235}{ORCID})

}
\keyword{internal}
