% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{oh_school_holidays_by_date}
\alias{oh_school_holidays_by_date}
\title{Get school holidays for a specific date for all countries}
\usage{
oh_school_holidays_by_date(date, language = "EN")
}
\arguments{
\item{date}{date (YYYY-MM-DD).}

\item{language}{Language ISO code (default: "EN").}
}
\value{
A data frame of school holidays for all countries.
}
\description{
Get school holidays for a specific date for all countries
}
\examples{
\dontrun{
oh_holiday_by_date(date = "2025-12-25")
}
}
