% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openholidaysR-package.R
\docType{package}
\name{openholidaysR-package}
\alias{openholidaysR}
\alias{openholidaysR-package}
\title{openholidaysR: Provides Access to the 'OpenHolidays' API}
\description{
Provides dates for public and school holidays for a number of countries and their subdivisions through the 'OpenHolidays' API at \url{https://www.openholidaysapi.org/en/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jfeldhege/openholidaysR}
  \item \url{http://www.johannesfeldhege.de/openholidaysR/}
  \item Report bugs at \url{https://github.com/jfeldhege/openholidaysR/issues}
}

}
\author{
\strong{Maintainer}: Johannes Feldhege \email{mail@johannesfeldhege.de} (\href{https://orcid.org/0000-0002-4665-5235}{ORCID}) [copyright holder]

}
\keyword{internal}
